# src/main/resources/openapi/openapi.yaml
openapi: 3.1.0

info:
  title: TromboneMaster API
  version: 1.0.0
  description: API for the TromboneMaster platform.

servers:
  - url: http://localhost:8080
    description: Local development

tags:
  - name: Auth
    description: Authentication and account registration
  - name: Users
    description: User management

paths:

  # ============================================================
  # Authentication
  # ============================================================

  /api/auth/register:

    post:
      tags:
        - Auth
      summary: Register a new user
      operationId: registerUser

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'

      responses:
        '201':
          description: User registered successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

        '400':
          description: Invalid registration data.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '409':
          description: Username or email already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'


  /api/auth/login:

    post:
      tags:
        - Auth
      summary: Authenticate a user
      operationId: login

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'

      responses:
        '200':
          description: Authentication successful.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'

        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '401':
          description: Invalid username or password.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'


  /api/auth/refresh:

    post:
      tags:
        - Auth
      summary: Refresh an access token
      operationId: refreshToken

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'

      responses:
        '200':
          description: Token refreshed successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'

        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '401':
          description: Invalid, expired, or revoked refresh token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'


  /api/auth/logout:

    post:
      tags:
        - Auth
      summary: Log out and revoke a refresh token
      operationId: logout

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'

      responses:
        '204':
          description: Logged out successfully.

        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '401':
          description: Invalid refresh token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'


  # ============================================================
  # Current authenticated user
  # ============================================================

  /api/users/me:

    get:
      tags:
        - Users
      summary: Get the current authenticated user
      operationId: getCurrentUser

      security:
        - bearerAuth: []

      responses:
        '200':
          description: Current user retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'


    put:
      tags:
        - Users
      summary: Update the current authenticated user
      operationId: updateCurrentUser

      security:
        - bearerAuth: []

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'

      responses:
        '200':
          description: Current user updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '409':
          description: Username or email already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'


    delete:
      tags:
        - Users
      summary: Delete the current authenticated user's account
      operationId: deleteCurrentUser

      security:
        - bearerAuth: []

      responses:
        '204':
          description: Account deleted successfully.

        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'


  # ============================================================
  # User administration
  # ============================================================

  /api/users:

    get:
      tags:
        - Users
      summary: Get users
      operationId: getUsers

      security:
        - bearerAuth: []

      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0

        - name: size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20

        - name: username
          in: query
          required: false
          schema:
            type: string
            minLength: 1

        - name: email
          in: query
          required: false
          schema:
            type: string
            format: email
            minLength: 3
            maxLength: 254

        - name: displayName
          in: query
          required: false
          schema:
            type: string
            minLength: 1

        - name: country
          in: query
          required: false
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
              minLength: 2
              maxLength: 2
              pattern: '^[A-Za-z]{2}$'

      responses:
        '200':
          description: Users retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPage'

        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '403':
          description: Access forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'


  /api/users/{userId}:

    parameters:
      - name: userId
        in: path
        required: true
        schema:
          type: string
          format: uuid

    get:
      tags:
        - Users
      summary: Get user
      operationId: getUser

      security:
        - bearerAuth: []

      responses:
        '200':
          description: User retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '403':
          description: Access forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'


    put:
      tags:
        - Users
      summary: Update user
      operationId: updateUser

      security:
        - bearerAuth: []

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'

      responses:
        '200':
          description: User updated successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

        '400':
          description: Invalid request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '403':
          description: Access forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '409':
          description: Username or email already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'


    delete:
      tags:
        - Users
      summary: Delete user
      operationId: deleteUser

      security:
        - bearerAuth: []

      responses:
        '204':
          description: User deleted successfully.

        '401':
          description: Authentication required.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '403':
          description: Access forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '404':
          description: User not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

        '500':
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'


components:

  # ============================================================
  # Security
  # ============================================================

  securitySchemes:

    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT


  # ============================================================
  # Schemas
  # ============================================================

  schemas:

    RegisterRequest:
      type: object
      required:
        - username
        - email
        - password
        - displayName
        - country
      properties:

        username:
          type: string
          minLength: 3
          maxLength: 30
          pattern: '^[A-Za-z][A-Za-z0-9_-]*$'
          example: peter_strele

        email:
          type: string
          format: email
          minLength: 3
          maxLength: 254
          example: peter@example.com

        password:
          type: string
          format: password
          minLength: 12
          maxLength: 128
          writeOnly: true
          example: Correct-Horse-Battery-Staple

        displayName:
          type: string
          minLength: 2
          maxLength: 30
          example: Peter Strele

        country:
          type: string
          minLength: 2
          maxLength: 2
          pattern: '^[A-Za-z]{2}$'
          example: PL


    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:

        username:
          type: string
          minLength: 1
          maxLength: 30
          example: peter_strele

        password:
          type: string
          format: password
          minLength: 1
          maxLength: 128
          example: SecurePassword123!


    LoginResponse:
      type: object
      required:
        - accessToken
        - refreshToken
        - tokenType
        - expiresIn
      properties:

        accessToken:
          type: string
          description: Short-lived JWT access token.

        refreshToken:
          type: string
          description: Refresh token used to obtain a new access token.

        tokenType:
          type: string
          example: Bearer

        expiresIn:
          type: integer
          format: int64
          description: Access token lifetime in seconds.
          example: 900


    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:

        refreshToken:
          type: string
          minLength: 1


    TokenResponse:
      type: object
      required:
        - accessToken
        - refreshToken
        - tokenType
        - expiresIn
      properties:

        accessToken:
          type: string
          description: Short-lived JWT access token.

        refreshToken:
          type: string
          description: Rotated refresh token.

        tokenType:
          type: string
          example: Bearer

        expiresIn:
          type: integer
          format: int64
          description: Access token lifetime in seconds.
          example: 900


    UpdateUserRequest:
      type: object
      required:
        - username
        - email
        - displayName
        - country
      properties:

        username:
          type: string
          minLength: 3
          maxLength: 30
          pattern: '^[A-Za-z][A-Za-z0-9_-]*$'
          example: trombonePeter

        email:
          type: string
          format: email
          minLength: 3
          maxLength: 254
          example: peter@example.com

        displayName:
          type: string
          minLength: 2
          maxLength: 30
          example: Peter

        country:
          type: string
          minLength: 2
          maxLength: 2
          pattern: '^[A-Za-z]{2}$'
          example: PL


    UserResponse:
      type: object
      required:
        - id
        - username
        - email
        - displayName
        - country
      properties:

        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000

        username:
          type: string
          example: trombonePeter

        email:
          type: string
          format: email
          example: peter@example.com

        displayName:
          type: string
          example: Peter

        country:
          type: string
          example: PL


    UserPage:
      type: object
      required:
        - content
        - page
        - size
        - totalElements
        - totalPages
      properties:

        content:
          type: array
          items:
            $ref: '#/components/schemas/UserResponse'

        page:
          type: integer
          example: 0

        size:
          type: integer
          example: 20

        totalElements:
          type: integer
          format: int64
          example: 42

        totalPages:
          type: integer
          example: 3


    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:

        code:
          type: string
          example: VALIDATION_ERROR

        message:
          type: string
          example: Request validation failed.

        errors:
          type: array
          items:
            $ref: '#/components/schemas/FieldError'


    FieldError:
      type: object
      required:
        - field
        - message
      properties:

        field:
          type: string
          example: username

        message:
          type: string
          example: size must be between 3 and 30